esther/backend_python/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic[email]
pydantic-settings
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator